requirements.txt
streamlit 
openai
memu-py
python-dotenv
pgvector